+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 # 1.02 (Not Released)
+- Fixed Hydrogen and Methane processing yielding a net gain of energy.
+- Fixed Player Detector constantly setting state changes despite no actual change.
 - Fixed Fusion casings using the wrong casings
 - Made the Quantum Chest give full stacks if the internal storage is above 64 when clicked.
 - Fixed gravel being constantly overwritten with sand on chunk load.
+- Fixed Emerald not having a macerator recipe into emerald dust
+- Removed the default thick reflector recipe for harder one
+- Fixed Rock Cutter defaulting to regular tool enchants in electric enchanter
+- Some more recipes I found would be useful in testing
 
 # 1.01
 - I recommend regenerating configs in this version.